% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_tools.R
\name{clean_refmap_prediction}
\alias{clean_refmap_prediction}
\title{clean_refmap_prediction}
\usage{
clean_refmap_prediction(x)
}
\arguments{
\item{x}{Seurat Object}
}
\value{
Seurat Object cleaned
}
\description{
clean reference mapping outputs
}
