% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_cellclassify.R
\name{convert_genedict}
\alias{convert_genedict}
\title{convert_genedict}
\usage{
convert_genedict(genelist, to = "human")
}
\arguments{
\item{genelist}{a list of gene symbols}

\item{to}{organism for the returned gene symbols, either "human" or "mouse"}
}
\description{
Convert a list of gene symbols between mouse MGI & human HGNC
}
