% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesymbol.R
\name{convert_human_to_mouse}
\alias{convert_human_to_mouse}
\title{convert_human_to_mouse}
\usage{
convert_human_to_mouse(x, unique = T)
}
\arguments{
\item{x}{A vector of HGNC gene symbols.}

\item{unique}{Whether to disable one-to-many MGI gene symbols to be returned. Defaults to {TRUE} to disable one-to-many mapping (TRUE)}
}
\value{
If {unique = T}; return a list of dataframe with "HGNC.symbol", "MGI.symbol" as human and corresponding mouse gene symbols. Otherwise return mouse gene symbols with one-to-one mapping only.
}
\description{
Convert human (HGNC) gene symbols to mouse (MGI) gene symbols; disable one-to-many mapping
}
\examples{
convert_mouse_to_human(c("MKI67", "TOP2A", "PCNA"), unique = T)

}
