% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesymbol.R
\name{convert_mouse_to_human}
\alias{convert_mouse_to_human}
\title{convert_mouse_to_human}
\usage{
convert_mouse_to_human(x, unique = T)
}
\arguments{
\item{x}{A vector of MGI gene symbols.}

\item{unique}{Whether to disable one-to-many HGNC gene symbols to be returned. Defaults to {TRUE} to disable one-to-many mapping}
}
\value{
If {unique = T}; return a list of dataframe with "MGI.symbol", "HGNC.symbol" as mouse and corresponding human gene symbols. Otherwise return human gene symbols with one-to-one mapping only.
}
\description{
Convert mouse (MGI) gene symbols to human (HGNC) gene symbols; disable one-to-many mapping
}
\examples{
convert_mouse_to_human(c("Mki67", "Top2a", "Pcna"), unique = T)

}
