% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_interoperability.R
\name{convert_seurat_to_anndata}
\alias{convert_seurat_to_anndata}
\title{convert_seurat_to_anndata}
\usage{
convert_seurat_to_anndata(
  x,
  h5ad,
  columns = NULL,
  reductions = c("umap", "pca"),
  snn = "RNA_snn",
  nn = "RNA_nn",
  assay = "RNA",
  adt_assay = NULL,
  return_genes = T,
  overwrite = F
)
}
\arguments{
\item{x}{Seurat object}

\item{h5ad}{path for output .h5ad file}

\item{columns}{a vector of metadata columns to keep. Defaults to NULL to keep all columns}

\item{assay}{assay name for gene expression. Defaults to "RNA"}

\item{adt_assay}{assay name for ADT counts if any}

\item{return_genes}{if TRUE, return genes from assays named c("CC", "BCR", "TCR", "MHC") to current assay. Defaults to TRUE}

\item{pca}{reduction name for X_pca. Defaults to "pca"}

\item{umap}{reduction name for X_umap. Defaults to "umap"}
}
\description{
Convert Seurat to AnnData with SeuratDisk::Convert()
}
