% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{dsb_normalize}
\alias{dsb_normalize}
\title{dsb_normalize}
\usage{
dsb_normalize(
  x,
  dir,
  denoise.counts = T,
  use.isotype.control = F,
  isotype.control.name.vec = NULL
)
}
\arguments{
\item{x}{a list of seurat objects}

\item{dir}{}
}
\value{
a list of seurat objects with DSB normalised "data" in ADT assay
}
\description{
Normalize CITEseq expressions by DSB normalization
}
