% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesymbol.R
\name{get_biomart}
\alias{get_biomart}
\title{get_biomart}
\usage{
get_biomart()
}
\value{
A list containing human and mouse mart objects.
}
\description{
Initializes the Ensembl Mart for human and mouse gene symbols.
}
