% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{get_palette}
\alias{get_palette}
\title{get_palette}
\usage{
get_palette(palette, n = 9)
}
\arguments{
\item{palette}{a vector of hexcode colors, or palettes from "RColorBrewer" or "viridis"}

\item{n}{number of colors}
}
\description{
get a vector of hexa color codes
}
