% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{integrate_v4}
\alias{integrate_v4}
\title{integrate_v4}
\usage{
integrate_v4(
  x,
  split.by,
  assay = "RNA",
  nfeatures = 3000,
  method = "rpca",
  k.filter = 200
)
}
\arguments{
\item{x}{Seurat object}

\item{split.by}{variable to split object by}

\item{assay}{assay name of gene expression. Defaults to "RNA"}

\item{nfeatures}{no. of features to select for integration}

\item{method}{integration method, see Seurat::indIntegrationAnchors()}

\item{k.filter}{k.filtered, see Seurat::indIntegrationAnchors(), reduce when no. of cells in samples are low.}
}
\description{
Wrapper function to integrate cells by samples
}
