% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_vdj.R
\name{plot_vdj}
\alias{plot_vdj}
\title{plot_vdj}
\usage{
plot_vdj(x, group.by, facet.by = NULL, variable = "isotype")
}
\arguments{
\item{x}{Seurat object}

\item{group.by}{column to group cells by}
}
\description{
plot vdj information for cells
}
