% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_vdj.R
\name{plot_vdj_qc}
\alias{plot_vdj_qc}
\title{plot_vdj_qc}
\usage{
plot_vdj_qc(x, group.by = "samples")
}
\arguments{
\item{x}{Seurat object}

\item{group.by}{column to group cells by}
}
\description{
plot proportion of cells that has VDJ library
}
