% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{process_seurat}
\alias{process_seurat}
\title{process_seurat}
\usage{
process_seurat(
  x,
  assay = "RNA",
  dims = 1:10,
  res = 0.4,
  reduction = NULL,
  harmony = F,
  group.by.vars = "samples"
)
}
\arguments{
\item{x}{Seurat object}

\item{assay}{assay name of gene expression. Defaults to "RNA"}

\item{dims}{no. of PCs to use. Defaults to 1:10}

\item{res}{clustering resolution. Defaults to 0.4}

\item{reduction}{reduction name if present. Defaults to NULL}

\item{harmony}{if TRUE, run Harmony integration for samples}

\item{group.by.vars}{variable to integrate by if harmony = TRUE}
}
\description{
Wrapper function to cluster cells by gene expression
}
