% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{qc_before_recluster}
\alias{qc_before_recluster}
\title{qc_before_recluster}
\usage{
qc_before_recluster(x, assay = "RNA")
}
\arguments{
\item{x}{Seurat object}

\item{assay}{assay name of gene expression. Defaults to "RNA"}
}
\description{
Subset cells under < 200 features after QC
}
