% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{remove_genes}
\alias{remove_genes}
\title{remove_genes}
\usage{
remove_genes(x, features = NULL, orig.assay = "RNA", new.assay)
}
\arguments{
\item{x}{Seurat object}

\item{features}{a vector genes/features to remove}

\item{orig.assay}{current assay name, defaults to "RNA"}

\item{new.assay}{new assay name to store removed genes}
}
\description{
Remove genes from current assay to new assay
}
