% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_scdblfinder}
\alias{run_scdblfinder}
\title{run_scdblfinder}
\usage{
run_scdblfinder(
  x,
  assay = "RNA",
  samples,
  truth = NULL,
  clusters,
  dbr,
  ncores = 1
)
}
\arguments{
\item{x}{Seurat object}

\item{assay}{assay name, defaults to "RNA"}

\item{truth}{whether cell demultiplexing (ground-truth) result is available.
Defaults to NULL, else set to metadata column with values of "Singlet" and "Doublet".}

\item{dbr}{doublet rate}

\item{ncores}{cores used for parallel processing, defaults to 1}

\item{cluster}{metadata column containing cluster labels}
}
\description{
predict doublets with scDblFinder
}
