% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scDensityUMAP}
\alias{scDensityUMAP}
\title{scDensityUMAP}
\usage{
scDensityUMAP(x, reduction = "umap", split.by, adjust = 1)
}
\arguments{
\item{x}{Seurat object}

\item{reduction}{reduction name, defaults to "umap"}

\item{split.by}{metadata column to group by}

\item{adjust}{adjust from stat_density_2d()}
}
\description{
Plot DensityUMAP from Seurat object
}
