% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scFeaturePlot}
\alias{scFeaturePlot}
\title{scFeaturePlot}
\usage{
scFeaturePlot(x, features, ncol, plot = F, ...)
}
\arguments{
\item{x}{Seurat object}

\item{features}{feature names stored in a vector}

\item{plot}{If TRUE, plot with cowplot::plot_grid(). Defaults to FALSE to return a list of ggplot objects}

\item{...}{arguments for Seurat::FeaturePlot()}

\item{assay}{assay name, defaults to "RNA"}

\item{slot}{slot name, defaults to "data"}
}
\description{
Plot FeaturePlot from Seurat object
}
