% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plotting.R
\name{scUMAP}
\alias{scUMAP}
\title{scUMAP}
\usage{
scUMAP(
  x,
  reduction = "umap",
  group.by,
  cols = NULL,
  count = F,
  pt.size = NULL,
  legend.ncol = 1,
  ...
)
}
\arguments{
\item{x}{Seurat object}

\item{reduction}{reduction name, defaults to "umap"}

\item{group.by}{metadata column to group by}

\item{cols}{a vector of colors, defaults to NULL}

\item{count}{If TRUE, count the no. of cells for each group. Defaults to FALSE}

\item{pt.size}{size of each cell/point on UMAP}

\item{legend.ncol}{no. of columns for group.by keys in legend}

\item{...}{arguments for Seurat::DimPlot()}
}
\description{
Plot UMAP from Seurat object
}
